move header sync validator to validation crate
