+++ base_layer/core/src/base_node/sync/header_sync/mod.rs
@@ -25,5 +25,3 @@ pub use error::BlockHeaderSyncError;
 
 mod synchronizer;
 pub use synchronizer::HeaderSynchronizer;
-
-mod validator;
