+++ base_layer/core/src/validation/header_sync_validator.rs
